+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.3.0
+
+* Delegated the signal handling to the signal-hook crate, so the same signal can
+  be shared with other things and the code in this crate is simplified (breaking
+  change in the `register_signal` method return type).
+* Fixed a bug with extra reopen just after creation.
+
 # 0.2.1
 
 * Lifted the many annoying limitations of `Handle::register_signal`.